New release for libsiu for FOAM 0.6.0 (31 commits)
Added gsl_vector_float reading, added check for empty lines.

Added debug prints

Added GSL to compilation flags for Csv class

Fixed Csv() constructor gsl_vector_float data.

Added test for Csv(gsl_vector_float*) constructor

Added simple_rand() function (used in FOAM)

Fixed help string.

Fixed include guards, added update() function

Re-arranged variables, made some variables private (use caller functions instead)

Bug in Gdk::GL::Config::create() ?

Added debug info on startup

Problem depens on OpenGL configuration, not on this code.

Improved thread id printing method

Method no longer platform dependent

Fixed intmax_t printing issue, using %ju instead of %lld

Fixed thread id printing issue, now platform independent

Separate utils.h from types.h

types.h had both datatypes and small functions, now separated.

Minor fix, std::string

Removed unused variable 'c'

Removed unused variable

Fix compiler warnings by explicit casting

Comment out unused variables

Made data members private & created get_ funcs

Added mutexholder with trylock() mechanism

mutexholdertry can be used to try to get a lock with trylock() in the
pthread++ framework.

Added proper CPPFLAGS for GSL

Flags were in CFLAGS first, which is not correct for -I. This didn't pose a problem until the libraries were in non-standard places.

Updated ignore file

Pattern '*-test' does not ignore 'io-test2', added * to the end.

Added performance logger library & test

This small library can record the time several processing steps take, and
uses a seperate thread to report this on a regular interval.

Fixed debugprint, added callbacks

Fixed DEBUGPRINT define, added separate function for reporting, added
sigc++ callback slot for users.

Update test using new features

Added perflogger to compilation targets

Added 'Update only on new data '

Using do_alwaysupdate, one can toggle whether updates will happen
always or only with new data.

Making memory allocation dynamic

PerfLog can log the performance of various stages of iterations, but
this needs to be stored somewhere. This branch automates memory allocation.

Implemented nstages, updated callback.

Callback is now (interval, nstages, minlat, maxlat, sumlat, avgcount)

Updated test for new PerfLog

Fix bug for slot_report() callback

More const, more docs, no format.h dependency

Applied more constness in functions when applicable. Improved code
description somewhat and incorporated some format.h code (vformat) to remove
the dependency on this file.

New test for Protocol with threads.

Added initializer

Added copyright statement

Added copyright statement

Added protocol-thread-test make directives

Included SigHandle in libsiu (from FOAM)

Added SigHandle make rules

define _GNU_SOURCE for strsignal(3) on Linux

Added <stdlib.h> for exit(3)

added max_quit_count

+++ src/pidfile.cc
@@ -42,7 +42,7 @@ pid_t pidfile::getpid() {
 	if(len == -1)
 		throw exception("Cannot read from pidfile '" + filename + "': " + strerror(errno));
 
-	if(len == 0 || len >= sizeof buf)
+	if(len == 0 || len >= (ssize_t) sizeof buf)
 		throw exception("Error parsing pid from '" + filename + "'");
 
 	return atoll(buf);
@@ -58,7 +58,7 @@ void pidfile::setpid() {
 	int len;
 	
 	len = snprintf(buf, sizeof buf, "%lli\n", (long long int)::getpid());
-	if(len <= 0 || len >= sizeof buf)
+	if(len <= 0 || len >= (ssize_t) sizeof buf)
 		throw exception("Error formatting pid");
 	
 	if(write(fd, buf, len) == -1)
